Finish cron job for sending email (not change email template yet)

+++ kalthailand-api-elife/src/main/java/th/co/krungthaiaxa/api/elife/data/PolicyNumberSetting.java
@@ -32,22 +32,22 @@ public class PolicyNumberSetting {
     @ElementEmail
     @ElementNotBlank
     private List<String> emailList;
-    
+
     /**
-     * The time trigger setting.
+     * The time trigger setting (in seconds).
      */
     @NotNull
-    private Integer timeTrigger;
+    private Long timeTrigger;
 
-    public Integer getTimeTrigger() {
-		return timeTrigger;
-	}
+    public Long getTimeTrigger() {
+        return timeTrigger;
+    }
 
-	public void setTimeTrigger(Integer timeTrigger) {
-		this.timeTrigger = timeTrigger;
-	}
+    public void setTimeTrigger(Long timeTrigger) {
+        this.timeTrigger = timeTrigger;
+    }
 
-	public Integer getTriggerPercent() {
+    public Integer getTriggerPercent() {
         return triggerPercent;
     }
 
@@ -64,12 +64,12 @@ public class PolicyNumberSetting {
     }
 
     @Override
-	public String toString() {
-		return "PolicyNumberSetting [id=" + id + ", triggerPercent=" + triggerPercent + ", emailList=" + emailList
-				+ ", timeTrigger=" + timeTrigger + "]";
-	}
+    public String toString() {
+        return "PolicyNumberSetting [id=" + id + ", triggerPercent=" + triggerPercent + ", emailList=" + emailList
+                + ", timeTrigger=" + timeTrigger + "]";
+    }
 
-	public String getId() {
+    public String getId() {
         return id;
     }
